# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elicta Service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/meetings": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Create Meeting Endpoint",
        "operationId": "create_meeting_endpoint_api_meetings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meetings/{meeting_id}/debrief/start": {
      "post": {
        "tags": [
          "debrief-session"
        ],
        "summary": "Start Debrief Session",
        "operationId": "start_debrief_session_api_meetings__meeting_id__debrief_start_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebriefConversationSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{meeting_id}/debrief/message": {
      "post": {
        "tags": [
          "debrief-session"
        ],
        "summary": "Send Message To Debrief Session",
        "operationId": "send_message_to_debrief_session_api_meetings__meeting_id__debrief_message_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebriefMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebriefConversationSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/settings": {
      "get": {
        "tags": [
          "admin-settings"
        ],
        "summary": "Get Settings",
        "description": "Current settings. Secrets appear as presence and a hint, never as values.",
        "operationId": "get_settings_api_admin_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin-settings"
        ],
        "summary": "Put Settings",
        "description": "Apply a settings change and return the new state.\n\nReturning the full settings rather than 204 lets the admin screen\nre-render from the service's view instead of its own optimistic one \u2014\nwhich matters most for secrets, where the client cannot know the\nresulting hint without being told.",
        "operationId": "put_settings_api_admin_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/settings/{key}/test": {
      "post": {
        "tags": [
          "admin-settings"
        ],
        "summary": "Test Connection",
        "description": "Check that a configured credential actually works.\n\nA 200 with `reachable: false` rather than an error status: the request\nitself succeeded, and the operator needs the reason rendered in the\nform next to the field, not an exception.",
        "operationId": "test_connection_api_admin_settings__key__test_post",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SecretKey"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionCheck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthMode": {
        "type": "string",
        "enum": [
          "api_key",
          "oauth_token"
        ],
        "title": "AuthMode",
        "description": "Which credential the service authenticates Claude calls with.\n\nBoth are \"bring your own\": an API key issued from the console, or an\nOAuth token from `claude setup-token` / `ant auth login`. Which one an\noperator has depends on how their organisation issues access, so this is\nan explicit choice rather than an inference from whichever field happens\nto be filled \u2014 an operator with both configured must be able to say which\none is live."
      },
...
      "ConnectionCheck": {
        "properties": {
          "key": {
            "$ref": "#/components/schemas/SecretKey"
          },
          "reachable": {
            "type": "boolean",
            "title": "Reachable"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "key",
          "reachable",
          "detail"
        ],
        "title": "ConnectionCheck",
        "description": "The result of testing a configured vendor credential.\n\n`detail` is a human-readable reason on failure. It is built from the\nvendor's error type, never from the credential."
      },
...
      "DebriefConversationSession": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "meeting_id": {
            "type": "string",
            "title": "Meeting Id"
          },
          "conversation_ref": {
            "type": "string",
            "title": "Conversation Ref"
          },
          "status": {
            "$ref": "#/components/schemas/DebriefSessionStatus"
          },
          "streaming_enabled": {
            "type": "boolean",
            "title": "Streaming Enabled"
          },
          "length_cap": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Cap"
          },
          "latency_budget_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Budget Seconds"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          },
          "nudge_dispositions": {
            "items": {
              "$ref": "#/components/schemas/NudgeDispositionRecord"
            },
            "type": "array",
            "title": "Nudge Dispositions"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/DebriefMessage"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "meeting_id",
          "conversation_ref",
          "status",
          "streaming_enabled",
          "length_cap",
          "latency_budget_seconds",
          "opened_at",
          "nudge_dispositions",
          "history"
        ],
        "title": "DebriefConversationSession",
        "description": "One meeting's open debrief conversation (PRD FR-7.1, FR-7.4).\n\n`conversation_ref` is the opaque handle the injected\n`OpenStreamingConversation` call returns for the underlying streaming\nconversation (backed by the Claude Agent SDK per the PRD) \u2014 this package\nnever inspects it, just carries it so a later message-exchange stage can\naddress the right conversation.\n\n`nudge_dispositions` is the inherited live-mode signal (FR-7.4): every\nnudge that fired during the call, each carrying whichever disposition \u2014\nfired, taken, or parked \u2014 it ended live mode with. It is persisted as\npart of the session itself rather than fetched separately, so the\ndebrief context always has that history the moment the session opens.\n\n`history` is every turn exchanged in the conversation so far, in order,\neach carrying its raw content blocks rather than extracted text (see\n`DebriefMessage`). It opens empty and grows one user turn plus one\nassistant turn per message sent."
      },
      "DebriefMessage": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/DebriefMessageRole"
          },
          "content": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Content"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content",
          "recorded_at"
        ],
        "title": "DebriefMessage",
        "description": "One turn of the debrief conversation, persisted verbatim.\n\n`content` is the underlying streaming conversation's own list of content\nblocks \u2014 for an assistant turn, the full `response.content` the Claude\nAgent SDK returns \u2014 rather than a text extraction of it. This package\nnever inspects the blocks, only carries them, since it has no way to\nknow which ones the vendor's server-side compaction has attached state\nto."
      },
      "DebriefMessageRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "DebriefMessageRequest",
        "description": "Request body for sending a free-form message into an open debrief conversation."
      },
      "DebriefMessageRole": {
        "type": "string",
        "enum": [
          "user",
          "assistant"
        ],
        "title": "DebriefMessageRole",
        "description": "Who authored one turn of the debrief conversation history."
      },
      "DebriefSessionStatus": {
        "type": "string",
        "enum": [
          "open"
        ],
        "title": "DebriefSessionStatus",
        "description": "Lifecycle state of a debrief conversation session.\n\nOnly `OPEN` is ever produced by this package \u2014 closing or resuming a\ndebrief conversation is a separate concern from opening one."
      },
...
      "EngagementContext": {
        "properties": {
          "client_organisation": {
            "type": "string",
            "title": "Client Organisation"
          },
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "commercial_context": {
            "type": "string",
            "title": "Commercial Context"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "scope_boundary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Boundary"
          },
          "target_requirements_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Requirements Template"
          }
        },
        "type": "object",
        "required": [
          "client_organisation",
          "sector",
          "commercial_context"
        ],
        "title": "EngagementContext",
        "description": "The engagement-level context a new meeting inherits automatically (PRD FR-3.1, FR-3.5, FR-3.7).\n\nMirrors the engagement's own context fields -- three captured at\nengagement creation (`client_organisation`, `sector`,\n`commercial_context`), three filled in later (`purpose`,\n`scope_boundary`, `target_requirements_template`) -- so a meeting-create\nresponse can echo them back for confirmation without this package\nimporting the engagement package's persistence-backed record type."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InferenceSettings": {
        "properties": {
          "auth_mode": {
            "$ref": "#/components/schemas/AuthMode",
            "description": "Whether calls authenticate with an API key or an OAuth token.",
            "default": "api_key"
          },
          "model": {
            "type": "string",
            "minLength": 1,
            "title": "Model",
            "description": "Model used for the context compiler and debrief pipeline.",
            "default": "claude-opus-5"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "Override the Anthropic API endpoint. Leave unset for the default."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "InferenceSettings",
        "description": "Non-secret inference configuration (architecture ADR-012, \u00a73.10)."
      },
...
      "MeetingCreateRequest": {
        "properties": {
          "engagement_id": {
            "type": "string",
            "minLength": 1,
            "title": "Engagement Id"
          },
          "capture_mode": {
            "type": "string",
            "minLength": 1,
            "title": "Capture Mode"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "engagement_id",
          "capture_mode"
        ],
        "title": "MeetingCreateRequest",
        "description": "What's specific to a new meeting -- its engagement, capture mode, and schedule (PRD FR-3.7, FR-3.8).\n\nDeliberately excludes every engagement-level context field: those are\ninherited automatically from `engagement_id` rather than re-entered\nhere. `capture_mode` is required since the `meetings` table declares it\n`NOT NULL` with no default; `scheduled_at` is optional since a meeting\ncan be created before a time is settled."
      },
      "MeetingCreateResponse": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "title": "Meeting Id"
          },
          "engagement_id": {
            "type": "string",
            "title": "Engagement Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "capture_mode": {
            "type": "string",
            "title": "Capture Mode"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "engagement_context": {
            "$ref": "#/components/schemas/EngagementContext"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "engagement_id",
          "state",
          "capture_mode",
          "engagement_context"
        ],
        "title": "MeetingCreateResponse",
        "description": "A newly created meeting, with its inherited engagement context attached for confirmation (PRD FR-3.7)."
      },
...
      "NudgeDisposition": {
        "type": "string",
        "enum": [
          "fired",
          "taken",
          "parked"
        ],
        "title": "NudgeDisposition",
        "description": "How one live-mode nudge was resolved by the time the call ended (PRD FR-7.4).\n\n`FIRED` is itself a disposition, not a placeholder for a missing one \u2014\nit means the nudge surfaced and the call ended before the operator took\nor parked it."
      },
      "NudgeDispositionRecord": {
        "properties": {
          "nudge_id": {
            "type": "string",
            "title": "Nudge Id"
          },
          "stub": {
            "type": "string",
            "title": "Stub"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "trigger_reason": {
            "type": "string",
            "title": "Trigger Reason"
          },
          "fired_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fired At"
          },
          "disposition": {
            "$ref": "#/components/schemas/NudgeDisposition"
          }
        },
        "type": "object",
        "required": [
          "nudge_id",
          "stub",
          "question",
          "trigger_reason",
          "fired_at",
          "disposition"
        ],
        "title": "NudgeDispositionRecord",
        "description": "One live-mode nudge as inherited into the debrief thread (PRD FR-7.4).\n\nMirrors the live nudge assistant's own surfaced nudge (a stub headline,\nthe full question, and the reason it fired \u2014 PRD FR-6.2/6.3/5.11) plus\nthe `disposition` it ended live mode with, so the debrief conversation\ncan discuss a specific nudge without a further round trip to the\nlive-mode store."
      },
...
      "SecretKey": {
        "type": "string",
        "enum": [
          "anthropic_api_key",
          "anthropic_oauth_token",
          "asr_vendor_api_key",
          "capture_vendor_api_key"
        ],
        "title": "SecretKey",
        "description": "The secrets an operator can configure.\n\nAn enum rather than free-form strings: a typo'd key name would otherwise\nsilently create a second, never-read secret while the real one stayed\nunset."
      },
      "SecretStatus": {
        "properties": {
          "key": {
            "$ref": "#/components/schemas/SecretKey"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint",
            "description": "Last four characters of the stored secret, to confirm which one is set."
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "key",
          "configured"
        ],
        "title": "SecretStatus",
        "description": "What a read of a secret returns: presence, not value."
      },
      "SecretUpdate": {
        "properties": {
          "key": {
            "$ref": "#/components/schemas/SecretKey"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "The secret, or an empty string to clear it."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "SecretUpdate",
        "description": "One secret being set or cleared.\n\nAn empty `value` clears the secret. That is the only way to remove one,\nand it is deliberately distinct from omitting the entry entirely, which\nleaves the stored secret alone."
      },
      "ServiceSettings": {
        "properties": {
          "inference": {
            "$ref": "#/components/schemas/InferenceSettings"
          },
          "vendors": {
            "$ref": "#/components/schemas/VendorSettings"
          },
          "secrets": {
            "items": {
              "$ref": "#/components/schemas/SecretStatus"
            },
            "type": "array",
            "title": "Secrets"
          },
          "durable": {
            "type": "boolean",
            "title": "Durable",
            "description": "Whether these settings survive a service restart. Surfaced so the admin screen can say so plainly, rather than letting an operator discover it after a restart.",
            "default": false
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ServiceSettings",
        "description": "Everything an operator can administer, with no secret values in it."
      },
...
      "SettingsUpdateRequest": {
        "properties": {
          "inference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferenceSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "vendors": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VendorSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "secrets": {
            "items": {
              "$ref": "#/components/schemas/SecretUpdate"
            },
            "type": "array",
            "title": "Secrets"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SettingsUpdateRequest",
        "description": "A settings save from the admin UI.\n\nEvery field is optional so the UI can save one section without having to\nresend the others \u2014 and, more importantly, so it never has to resend a\nsecret it does not have."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VendorSettings": {
        "properties": {
          "asr_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asr Base Url"
          },
          "capture_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capture Base Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "VendorSettings",
        "description": "Non-secret endpoints for the speech vendors (ADR-004, ADR-011)."
      },
...
    }
  }
}